// web/src/types.ts
import { z } from 'zod';

export const MODERATION_CATEGORIES = [
  'OFFENSIVE',
  'OFF_BRAND',
  'VIOLENCE',
  'NONE',
] as const;

export type ModerationCategory = (typeof MODERATION_CATEGORIES)[number];
export const ModerationCategoryZod = z.enum([...MODERATION_CATEGORIES]);

export type SessionStatus = 'DISCONNECTED' | 'CONNECTING' | 'CONNECTED';

export interface ToolParameterProperty {
  type: string;
  description?: string;
  enum?: string[];
  pattern?: string;
  properties?: Record<string, ToolParameterProperty>;
  required?: string[];
  additionalProperties?: boolean;
  items?: ToolParameterProperty;
}

export interface ToolParameters {
  type: string;
  properties: Record<string, ToolParameterProperty>;
  required?: string[];
  additionalProperties?: boolean;
}

export interface Tool {
  type: 'function';
  name: string;
  description: string;
  parameters: ToolParameters;
}

export interface AgentConfig {
  name: string;
  publicDescription: string;
  instructions: string;
  tools: Tool[];
  toolLogic?: Record<
    string,
    (
      args: any,
      transcriptLogsFiltered: TranscriptItem[],
      addTranscriptBreadcrumb?: (title: string, data?: any) => void
    ) => Promise<any> | any
  >;
  downstreamAgents?:
    | AgentConfig[]
    | { name: string; publicDescription: string }[];
}

export type AllAgentConfigsType = Record<string, AgentConfig[]>;

export interface GuardrailResultType {
  status: 'IN_PROGRESS' | 'DONE';
  testText?: string;
  category?: ModerationCategory;
  rationale?: string;
}

export interface TranscriptItem {
  itemId: string;
  type: 'MESSAGE' | 'BREADCRUMB';
  role?: 'user' | 'assistant';
  title?: string;
  data?: Record<string, any>;
  expanded: boolean;
  timestamp: string;
  createdAtMs: number;
  status: 'IN_PROGRESS' | 'DONE';
  isHidden: boolean;
  guardrailResult?: GuardrailResultType;
}

export interface Log {
  id: number;
  timestamp: string;
  direction: string;
  eventName: string;
  data: any;
  expanded: boolean;
  type: string;
}

export interface ServerEvent {
  type: string;
  event_id?: string;
  item_id?: string;
  transcript?: string;
  delta?: string;
  session?: { id?: string };
  item?: {
    id?: string;
    object?: string;
    type?: string;
    status?: string;
    name?: string;
    arguments?: string;
    role?: 'user' | 'assistant';
    content?: { type?: string; transcript?: string | null; text?: string }[];
  };
  response?: {
    output?: {
      id: string;
      type?: string;
      name?: string;
      arguments?: any;
      call_id?: string;
      role: string;
      content?: any;
    }[];
    metadata: Record<string, any>;
    status_details?: { error?: any };
  };
}

export interface LoggedEvent {
  id: number;
  direction: 'client' | 'server';
  expanded: boolean;
  timestamp: string;
  eventName: string;
  eventData: Record<string, any>;
}

export const GuardrailOutputZod = z.object({
  moderationRationale: z.string(),
  moderationCategory: ModerationCategoryZod,
  testText: z.string().optional(),
});

export type GuardrailOutput = z.infer<typeof GuardrailOutputZod>;